Store user profile - Abstract logic for managing user session to AuthManager.kt. - Store access token in AuthState.kt - Store user phone number in database. - Add UUID serializer to network requests and responses.

+++ app/src/main/kotlin/com/anshtya/jetx/profile/data/AvatarManager.kt
@@ -14,6 +14,13 @@ import java.io.FileOutputStream
 import javax.inject.Inject
 import javax.inject.Singleton
 
+/**
+ * Class for managing user avatar operations.
+ *
+ * This class provides functionality for retrieving, saving, and updating
+ * user avatars. Avatars are stored securely in the app’s private storage,
+ * ensuring that user data is isolated from external access.
+ */
 @Singleton
 class AvatarManager @Inject constructor(
     @ApplicationContext private val context: Context,
@@ -39,18 +46,29 @@ class AvatarManager @Inject constructor(
                 ext = "jpg"
             )
             ensureActive()
-            FileOutputStream(file).use { outputStream ->
+            FileOutputStream(file, false).use { outputStream ->
                 outputStream.write(byteArray)
             }
-            
+
             file
+        }
+    }
+
+    suspend fun saveTempAvatar(
+        userId: String,
+        byteArray: ByteArray
+    ): Result<File> = runCatching {
+        withContext(ioDispatcher) {
+            val tempFile = File.createTempFile(userId, ".jpg", context.cacheDir)
+            tempFile.writeBytes(byteArray)
+            tempFile
         }
     }
 
     suspend fun updateAvatar(
         userId: String,
         newByteArray: ByteArray
-    ) {
+    ): Result<File> = runCatching {
         withContext(ioDispatcher) {
             val avatarDirectory = FileUtil.getAvatarDirectory(context)
             val avatarFile = File(avatarDirectory, userId)
@@ -58,17 +76,21 @@ class AvatarManager @Inject constructor(
             FileOutputStream(avatarFile, false).use { outputStream ->
                 outputStream.write(newByteArray)
             }
+
+            avatarFile
         }
     }
 
     suspend fun deleteAvatar(
         userId: String
-    ) {
+    ): Result<Unit> = runCatching {
         withContext(ioDispatcher) {
             val avatarDirectory = FileUtil.getAvatarDirectory(context)
             val avatarFile = File(avatarDirectory, userId)
             ensureActive()
-            avatarFile.delete()
+            if (avatarFile.exists()){
+                avatarFile.delete()
+            }
         }
     }
 }
